refactor(courses): restructure courses module; remove unused service and controller
