--- backend/src/courses/courses.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('courses')
-export class CoursesController {}
